requirements.txt
black==23.7.0
cfgv==3.3.1
click==8.1.6
distlib==0.3.7
exceptiongroup==1.1.2
filelock==3.12.2
identify==2.5.26
iniconfig==2.0.0
mypy==1.4.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
progressbar2==4.2.0
pytest==7.4.0
python-utils==3.7.0
PyYAML==6.0.1
tomli==2.0.1
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
virtualenv==20.24.2
